master

+++ src/main/java/tn/esprit/spring/service/DepartmentService.java
@@ -10,11 +10,11 @@ public interface DepartmentService {
 
 	Department addDepartment(Department D);
 
-	void deleteDepartment(Long idDepartment);
+	void deleteDepartment(int idDepartment);
 
 	Department updateDepartment(Department D);
 
-	Department retrieveDepartment(Long idDepartment);
+	Department retrieveDepartment(int idDepartment);
 
 	List<Department> retrieveAllDepartments();
